move leastsqfitter to kit. (resolves #616)
